Add files via upload
This lab will have you become familiar with the functionality of a stack.

You will solve two problems we’ve discussed in class–testing for palindromes and balanced parenthesis.

You will create two methods, isPalindrome and isBalanced. These methods should both return a boolean.

Your input will be 4 lines to test isPalindrome and 4 lines to test isBalanced. These 8 lines will be followed by a blank line. Remember you can write print statements like System.out.println(methodThatReturnsSomething(in.nextLine()));

Refer to the slides for testing if something is a palindrome. A palindrome is a word that reads the same forwards and backwards. You must use a stack to test whether a string is a palindrome. For example, racecar is a palindrome but dog is not. Your input will be all lowercase.

For checking balanced parenthesis your input will only contain opening and closing parenthesis – ( and ) – you must use a stack to check if you are given a balanced expression–contains an equal number of opening and closing parenthesis. For example, (()) and ()() are balanced while ()( and (() are not. We discussed you can push for an opening parenthesis and pop for a closing parenthesis. You may have stack errors with this for unbalanced parenthesis. Remember you can do checks like this: