Add files via upload
For this program you will implement a hospital triage system using a priority queue.
Create a java file named Program3. This will be the name of the public class.
Create a class Triage that implements the Comparable interface. It has three fields, name, problem, and severity.

This program will run through a menu. The menu should be structured as below:

1. Intake patient  
2. Call patient in
3. How many patients waiting  
4. Patient gave up and left  
5. Print Options  
6. End program
Option 1 asks for the patient name, problem, and severity. Severity options are mild, moderate, and severe. This severity field will be what you base priority on. Think about what data type would be best to store severity as in the object and what would make these easiest to compare by in the compareTo method.

Option 2 calls the patient with the highest severity. The severity order is: 1. Severe, 2. Moderate, 3. Mild. Checks if there are no patients left.

Option 3 lists how many patients are remaining.

Option 4 asks for a patient name and then removes them from the queue. Check if the patient exists or not. You can iterate over a priority queue like a list.

Option 5 prints the menu.

Option 6 ends the program.